effects: move ShardGun and RocketGun
