+++ src/game/effects/gun.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "global/types.h"
+
+int16_t Effect_ShardGun(
+    int32_t x, int32_t y, int32_t z, int16_t speed, PHD_ANGLE y_rot,
+    int16_t room_num);
+
+int16_t Effect_RocketGun(
+    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
+    int16_t room_num);
