+++ src/game/objects/ai/natla.c
@@ -7,7 +7,7 @@
 #include "game/items.h"
 #include "game/lot.h"
 #include "game/music.h"
-#include "game/objects/effects/missile.h"
+#include "game/effects/gun.h"
 #include "game/people.h"
 #include "game/random.h"
 #include "game/sound.h"
@@ -103,7 +103,8 @@ void NatlaControl(int16_t item_num)
                 item->goal_anim_state = NATLA_RUN;
             }
             if (timer >= 20) {
-                int16_t fx_num = CreatureEffect(item, &g_NatlaGun, ShardGun);
+                int16_t fx_num =
+                    CreatureEffect(item, &g_NatlaGun, Effect_ShardGun);
                 if (fx_num != NO_ITEM) {
                     FX_INFO *fx = &g_Effects[fx_num];
                     gun = fx->pos.x_rot;
@@ -116,7 +117,8 @@ void NatlaControl(int16_t item_num)
         case NATLA_RUN:
             tilt = angle;
             if (timer >= 20) {
-                int16_t fx_num = CreatureEffect(item, &g_NatlaGun, ShardGun);
+                int16_t fx_num =
+                    CreatureEffect(item, &g_NatlaGun, Effect_ShardGun);
                 if (fx_num != NO_ITEM) {
                     FX_INFO *fx = &g_Effects[fx_num];
                     gun = fx->pos.x_rot;
@@ -222,7 +224,8 @@ void NatlaControl(int16_t item_num)
                 item->goal_anim_state = NATLA_STOP;
             }
             if (timer >= 30) {
-                int16_t fx_num = CreatureEffect(item, &g_NatlaGun, RocketGun);
+                int16_t fx_num =
+                    CreatureEffect(item, &g_NatlaGun, Effect_RocketGun);
                 if (fx_num != NO_ITEM) {
                     FX_INFO *fx = &g_Effects[fx_num];
                     gun = fx->pos.x_rot;
@@ -244,17 +247,18 @@ void NatlaControl(int16_t item_num)
 
         case NATLA_SHOOT:
             if (!item->required_anim_state) {
-                int16_t fx_num = CreatureEffect(item, &g_NatlaGun, RocketGun);
+                int16_t fx_num =
+                    CreatureEffect(item, &g_NatlaGun, Effect_RocketGun);
                 if (fx_num != NO_ITEM) {
                     FX_INFO *fx = &g_Effects[fx_num];
                     gun = fx->pos.x_rot;
                 }
-                fx_num = CreatureEffect(item, &g_NatlaGun, RocketGun);
+                fx_num = CreatureEffect(item, &g_NatlaGun, Effect_RocketGun);
                 if (fx_num != NO_ITEM) {
                     FX_INFO *fx = &g_Effects[fx_num];
                     fx->pos.y_rot += (Random_GetControl() - 0x4000) / 4;
                 }
-                fx_num = CreatureEffect(item, &g_NatlaGun, RocketGun);
+                fx_num = CreatureEffect(item, &g_NatlaGun, Effect_RocketGun);
                 if (fx_num != NO_ITEM) {
                     FX_INFO *fx = &g_Effects[fx_num];
                     fx->pos.y_rot += (Random_GetControl() - 0x4000) / 4;
